% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_fold.R
\name{create_fold}
\alias{create_fold}
\title{Create folds for model validation}
\usage{
create_fold(the_data, cluster, num_folds)
}
\arguments{
\item{the_data}{the data}

\item{cluster}{a string of the variable name representing the cluster-level}

\item{num_folds}{the number of folds to be used (k = 2 = split-half validation, k > 2 = k-fold cross-validation)}
}
\value{
the_data with fold as a new variable
}
\description{
Appends a new variable, fold, for the purposes of being used with validate_ml. This must be done as an initial step in order
to allow for comparisons between models, so repeated calls to validate_ml with different statistical models will be using the same
split structure. Folds are created at the cluster-level.
}
\examples{
data_fold <- create_fold(HSB_data, "School", 2)
}
\references{
Martin, D. P. (2015). Efficiently exploring multilevel data with recursive partitioning (Unpublished doctoral
dissertation). University of Virginia, Charlottesville, VA.
}

