% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HSB_data.R
\docType{data}
\name{HSB_data}
\alias{HSB_data}
\title{High School and Beyond - 1982}
\format{A dataframe with 7,185 observations of 11 variables}
\usage{
data(HSB_data)
}
\description{
The data are a subsample of the 1982 High School and Beyond survey, with 7,185 students from 160
schools (range = 14 - 67).
}
\details{
\itemize{
  \item School: school id
  \item Minority: minority status (1 = minority, 0 = non-minority)
  \item Sex: sex of the child (1 = female, 0 = male)
  \item SES: standardized scale of socio-economic status constructed from variables measuring parental education,
  occupation, and income)
  \item MathAch: math achievement
  \item Size: number of students per school
  \item Sector: sector of school (1 = Catholic, 0 = public)
  \item PRACAD: proportion of students on the academic track in a given school
  \item DISCLIM: a scale measuring disciplinary climate
  \item HIMNTY: whether a school has a high proportion of minority students (1 = more than 40\% minority enrollment,
  0 = less than 40\%)
  \item MEANSES: mean of the SES values of students in a given school
}
}
\keyword{High}
\keyword{School}

