% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importance_ml.R
\name{importance_ml}
\alias{importance_ml}
\title{Plot variable importance for multilevel/forest models}
\usage{
importance_ml(mod_list)
}
\arguments{
\item{mod_list}{list of the models to be compared and plotted}
}
\value{
Plots variable importance for the models given
}
\description{
Plots and compares variable importances for an arbitrary number of statistical models (limited to randomForest, cforest, and lme4
model objects). Calculates variable importance for forest models using the randomForest and cforest built-in procedures using
permutation tests. Variable importance for lme4 models are naively defined as the p-value of the respective parameter, estimated
via a Satterthwaite approximation from the lmerTest package. Be sure to build your lme4 model with the lmerTest package loaded
or else an error will be displayed. Variable importances are standardized to allow for easier comparisons.
Plots are ordered such that the most important variable (averaged across all models) is the furthest left on the x-axis.
}
\examples{
## Not run:

# Run three models (rf, cf, and lmer)

rf_mod <- randomForest(MathAch ~ Minority + Sex + SES + Size +
                       Sector + PRACAD + DISCLIM + HIMINTY + MEANSES,
                       data = HSB_data)

cf_mod <- cforest(MathAch ~ Minority + Sex + SES + Size +
                  Sector + PRACAD + DISCLIM + HIMINTY + MEANSES,
                  data = HSB_data)

lmer_mod <- lmer(MathAch ~ Minority + Sex + SES + Size +
                 Sector + PRACAD + DISCLIM + HIMINTY +
                 MEANSES + (1 | School),
                 data = HSB_data)

# Plot variable importance comparison (use explicit names in the list for labelling purposes on the plot if desired)

importance_ml(list(rf = rf_mod, cf = cf_mod, lmer = lmer_mod))

## End(Not run)
}
\references{
Martin, D. P. (2015). Efficiently exploring multilevel data with recursive partitioning (Unpublished doctoral
dissertation). University of Virginia, Charlottesville, VA.
}

