\name{mleda}
\alias{mleda}
\docType{package}
\title{
Multilevel Exploratory Data Analysis
}
\description{
Perform efficient exploratory data analysis for multilevel data 
structures (limited to two levels) using random forests and 
mixed-effects models. Uses quick-and-dirty partial dependence 
plots for visualization (see plotmo package).
}
\details{
\tabular{ll}{
Package: \tab mleda\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-20\cr
License: \tab Apache License 2.0\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Daniel Martin <dpmartin42@gmail.com>
}
\references{
Dissertation reference here. 
}
\keyword{multilevel random forest, exploratory data analysis}
\seealso{
}
\examples{
~~ simple examples of the most important functions ~~
}
