% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_ml.R
\name{plot_ml}
\alias{plot_ml}
\title{Exploratory plots for multilevel data}
\usage{
plot_ml(the_data, the_mod, var_name, var_level, cluster, outcome,
  interact = FALSE, reference)
}
\arguments{
\item{the_data}{the dataset to be used}

\item{the_mod}{optional model that is either a randomForest, cforest, or lme4 model object
(omit this argument to use loess curves on raw data)}

\item{var_name}{vector of strings of variable names to plot}

\item{var_level}{vector of numbers indicating levels (1 or 2 only)}

\item{cluster}{string of variable name that represents cluster}

\item{outcome}{string of variable name that represents the outcome}

\item{interact}{boolean indicating if an interaction should be plotted (limited to two variables)}

\item{reference}{reference category in the case of classification, with the original reference category as the default}
}
\value{
Plots requested main effects/interactions
}
\description{
Quickly plot main effects and two-way interactions for multilevel data structures (limited to two levels) to
visualize the relations between a set of variables and a given outcome. Uses either raw data (loess curves)
or a "poor-man's" partial dependence plot from a randomForest, cforest, or lme4 model object.
}
\examples{
## Not run:

# Plot raw data for SES, MEANSES, and an interaction

plot_ml(the_data = HSB_data, outcome = "MathAch",
        var_name = c("SES", "MEANSES"),
        var_level = c(1, 2), cluster = "School",
        interact = TRUE)

# Random forest example

rf_mod <- randomForest(MathAch ~ Minority + Sex + SES + Size +
                       Sector + PRACAD + DISCLIM + HIMINTY + MEANSES,
                       data = HSB_data)

plot_ml(the_data = HSB_data, outcome = "MathAch",
        var_name = c("SES", "MEANSES"),
        var_level = c(1, 2), cluster = "School",
        interact = TRUE, the_mod = rf_mod)

# Conditional inference forest example

cf_mod <- cforest(MathAch ~ Minority + Sex + SES + Size +
                  Sector + PRACAD + DISCLIM + HIMINTY + MEANSES,
                  data = HSB_data)

plot_ml(the_data = HSB_data, outcome = "MathAch",
        var_name = c("SES", "MEANSES"),
        var_level = c(1, 2), cluster = "School",
        interact = TRUE, the_mod = cf_mod)

# Multi-level model example

lmer_mod <- lmer(MathAch ~ Minority + Sex + SES + Size +
                 Sector + PRACAD + DISCLIM + HIMINTY +
                 MEANSES + (1 | School),
                 data = HSB_data)

plot_ml(the_data = HSB_data, outcome = "MathAch",
        var_name = c("SES", "MEANSES"),
        var_level = c(1, 2), cluster = "School",
        interact = TRUE, the_mod = lmer_mod)

## End(Not run)
}
\references{
Martin, D. P. (2015). Efficiently exploring multilevel data with recursive partitioning (Unpublished doctoral
dissertation). University of Virginia, Charlottesville, VA.
}

