% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mleda.R
\name{plot_ml}
\alias{plot_ml}
\title{plots stuff}
\usage{
plot_ml(the_data, the_mod, var_name, var_level, cluster, type = "raw",
  interact = FALSE, reference)
}
\arguments{
\item{the_data}{- the dataset to be used}

\item{the_mod}{- the model (limited to randomForest objects right now)}

\item{var_name}{- vector of strings of variable names to plot}

\item{var_level}{- vector of numbers indicating levels (1 or 2 only)}

\item{cluster}{- string of variable name that represents cluster}

\item{type}{- plot raw or predicted values}

\item{interact}{- boolean indicating if an interaction should be plotted (limited to two variables)}

\item{reference}{- reference category in the case of classification. Default is most endorsed category}
}
\value{
plot of main effects/interactions
}
\description{
plots stuff
}
\examples{
# to do
}
\references{
[1] How to add numbers
}

