% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate_ml.R
\name{validate_ml}
\alias{validate_ml}
\title{validates stuff}
\usage{
validate_ml(the_data, formula, stat_method, valid_method, cluster)
}
\arguments{
\item{the_data}{- the dataset to be used}

\item{formula}{- model formula for either a randomForest, cforest, or lme4 model object}

\item{stat_method}{- statistical method to be used (rf for randomForest, cf for cforest, hlm for multilevel model)}

\item{valid_method}{- validation method of either split-half validation or 5-fold cross-validation (splitting occurs at the cluster-level)}

\item{cluster}{- string of a variable name that represents the cluster-level}
}
\value{
prints proportion of variation accounted for in the outcome
}
\description{
validates stuff
}
\examples{
# to do
}
\references{
[1] foo
}

